change coupon top bar logic
